Update accessibility tests
