+++ __tests__/a11y/example.spec.js
@@ -1,7 +1,5 @@
 import { test, expect } from '@playwright/test';
-import AxeBuilder from '@axe-core/playwright';
-import fs from 'fs';
-import path from 'path';
+import { analyzePage } from './shared/axeHelper.js';
 
 test.describe('Accessibility checks', () => {
   test('Home page should have no accessibility violations', async ({ page }) => {
@@ -9,21 +7,9 @@ test.describe('Accessibility checks', () => {
     await page.goto('http://localhost:3000'); // Replace with your application's URL
 
     // Perform accessibility analysis
-    const accessibilityScanResults = await new AxeBuilder({ page }).analyze();
+    const accessibilityScanResults = await analyzePage(page);
 
-    // Define the output directory and file path
-    const outputDir = path.resolve(process.cwd(), 'a11y-results');
-    const filePath = path.join(outputDir, 'home.json');
-
-    // Ensure the output directory exists
-    if (!fs.existsSync(outputDir)) {
-      fs.mkdirSync(outputDir);
-    }
-
-    // Write the accessibility scan results to a JSON file
-    fs.writeFileSync(filePath, JSON.stringify(accessibilityScanResults, null, 2));
-    console.log(`Accessibility scan results written to ${filePath}`);
-
+    console.log('Accessibility scan results:\n\n', accessibilityScanResults);
     // Assert that there are no accessibility violations
     expect(accessibilityScanResults.violations).toEqual([]);
   });
